Del contact page & Add new routes
