+++ src/types/data.ts
@@ -1,0 +1,10 @@
+export type searchByNameDataType = {
+  borders?: string[],
+  capital?: string[],
+  continents?: string,
+  currencies?: Object,
+  flags?: { svg: string }
+  languages?: Object,
+  name?: { common: string },
+  population?: number
+}
